requirements.txt
Django==4.2.17
django-ckeditor==6.7.0
djangorestframework==3.14.0
django-redis==5.4.0
redis==5.0.1